c-modularization
